requirements.txt
pandas
numpy
biopython
seaborn
matplotlib
scipy
